landing-page-backend/pyproject.toml
[project]
name = "landing-page-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.2.1",
]
